+++ modules/features/widgets/src/main/kotlin/au/com/shiftyjelly/pocketcasts/widget/ui/SmallPlayer.kt
@@ -7,14 +7,10 @@ import androidx.compose.ui.unit.min
 import androidx.glance.GlanceModifier
 import androidx.glance.LocalContext
 import androidx.glance.LocalSize
-import androidx.glance.action.clickable
 import androidx.glance.layout.Alignment
 import androidx.glance.layout.Box
 import androidx.glance.layout.fillMaxSize
 import androidx.glance.layout.size
-import androidx.glance.layout.width
-import androidx.glance.semantics.contentDescription
-import androidx.glance.semantics.semantics
 import au.com.shiftyjelly.pocketcasts.widget.action.OpenPocketCastsAction
 import au.com.shiftyjelly.pocketcasts.widget.action.controlPlaybackAction
 import au.com.shiftyjelly.pocketcasts.widget.data.LocalSource
@@ -30,8 +26,8 @@ internal fun SmallPlayer(state: SmallPlayerWidgetState) {
     }
     val contentDescription = when {
         state.episode == null -> LR.string.pocket_casts
-        state.isPlaying -> LR.string.play_episode
-        else -> LR.string.pause_episode
+        // don't use a pause state as TalkBack will take focus and cause the state to change
+        else -> LR.string.play_episode
     }.let { LocalContext.current.getString(it) }
 
     val size = min(LocalSize.current.width, LocalSize.current.height)
@@ -39,12 +35,7 @@ internal fun SmallPlayer(state: SmallPlayerWidgetState) {
     WidgetTheme(state.useDynamicColors) {
         Box(
             contentAlignment = Alignment.Center,
-            modifier = GlanceModifier
-                .fillMaxSize()
-                .clickable(controlPlayback)
-                .semantics {
-                    this.contentDescription = contentDescription
-                },
+            modifier = GlanceModifier.fillMaxSize(),
         ) {
             EpisodeImage(
                 episode = state.episode,
@@ -52,12 +43,14 @@ internal fun SmallPlayer(state: SmallPlayerWidgetState) {
                 size = size,
                 backgroundColor = { it.background },
                 onClick = controlPlayback,
+                contentDescription = contentDescription,
             )
             if (state.episode != null) {
                 val buttonSize = (size / 2.3f).coerceAtMost(48.dp)
                 PlaybackButton(
                     isPlaying = state.isPlaying,
                     iconPadding = buttonSize / 4.75f,
+                    isClickable = false,
                     modifier = GlanceModifier.size(buttonSize),
                 )
             }