+++ CHANGELOG.md
@@ -21,6 +21,8 @@
         ([#3769](https://github.com/Automattic/pocket-casts-android/pull/3769))
     *   Reorder podcasts sorting order based on their popularity. 
         ([#3733](https://github.com/Automattic/pocket-casts-android/pull/3733))
+    *   Improve small widget accessibility.
+        ([#3770](https://github.com/Automattic/pocket-casts-android/pull/3770))
 
 7.84
 -----